### 2024-10-06 9a226c7
update the scipy simps to simpson with checks, update doc
