### 2024-10-06 0abb218
udpate v number to 0.3.2.b
