### 2024-11-18 #7493 @editorialbot
[REVIEW]: ysoisochrone: A Python package to estimate masses and ages for YSOs
**Submitting author:** @DingshanDeng  (<a href="http://orcid.org/0000-0003-0777-7392">Dingshan Deng</a>)
**Repository:** https://github.com/DingshanDeng/ysoisochrone
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
