### 2024-12-05 #7493 @warrickball
Hi @HeloiseS & @vsquicciarini, I just wanted to check in on how your reviews were progressing and if you need anything from me. Note that we find JOSS reviews are most efficient if you start commenting on the submission as you work through the checklist, rather than waiting to accumulate a single, monolithic review, as one would for a traditional journal.
