### 2024-12-11 DingshanDeng/ysoisochrone#2 @HeloiseS
[JOSS] - Install instructions
This is an issue relating to your submission to JOSS (see [JOSS Issue 7493](https://github.com/openjournals/joss-reviews/issues/7493)).

Although your install instruction technically works, there are a couple of improvements I would like to request:

- [ ] Instead of downloading the .zip file it is more common and practical to ask people to clone the github repo.
- [ ] Have you considered actually uploading your software to pip to make it possible to run: `pip install ysoisochrone` ?
- [ ] As an extended (optional) suggestion, to make your life easier, you can set up github actions to automatically upload a new version of your software to pipy when a new tag is pushed. (see [instructions](https://packaging.python.org/en/latest/guides/publishing-package-distribution-releases-using-github-actions-ci-cd-workflows/) )
