### 2024-12-11 DingshanDeng/ysoisochrone#3 @HeloiseS
[JOSS] - Unit tests
This is an issue relating to your submission to JOSS (see https://github.com/openjournals/joss-reviews/issues/7493).

I notice you don't have a suit of unit tests.
Testing being a JOSS requirement I think that is the first thing you should start working on as it will probably take a bit of time to write those.
I can recommend [pytest](https://docs.pytest.org/en/stable/how-to/unittest.html) as I've been using it for years, it's simple the get started and it has a lot of extra functionalities if you want to be fancy (like testing grids of parameters in the inputs of your functions).

In addition, I would recommend including a github workflow to run your tests automatically and a variety of environments when you push or create a pull request.
For a very bare bones example of such a workflow you can check [this one out](https://github.com/HeloiseS/atlasapiclient/blob/main/.github/workflows/testing.yml).
