### 2024-12-11 DingshanDeng/ysoisochrone#4 @HeloiseS
[JOSS] - Community Guidelines
This is an issue relating to your submission to JOSS (see https://github.com/openjournals/joss-reviews/issues/7493).

Another key requirement of JOSS is clear community guidelines:

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Please add those to your README so i can tick the box ;)
