### 2024-12-13 DingshanDeng/ysoisochrone#5 @HeloiseS
[JOSS] - numpy not in setup.py 
This is an issue relating to your submission to JOSS (see https://github.com/openjournals/joss-reviews/issues/7493).

- [ ] Add numpy to the conf.py

Although numpy is in the requirements.txt it's not in the conf.py, which means that when i installed the package with the instructions provided and started running the notebooks numpy was not installed.
