### 2024-12-13 DingshanDeng/ysoisochrone#5 @HeloiseS
Just to confirm it was a me problem ;) install worked
