### 2024-12-13 DingshanDeng/ysoisochrone#6 @HeloiseS
[JOSS] - Tutorial1 Quick Start  - TypeError when running derive_stellar_mass_age
This is an issue relating to your submission to JOSS (see openjournals/joss-reviews#7493).

When running the cell
```
best_logmass_output, best_logage_output, lmass_all, lage_all =\
    ysoisochrone.bayesian.derive_stellar_mass_age(df_prop, model='Baraffe_n_Feiden', plot=True)
```

I get
```
Cell In[6], line 2
      1 best_logmass_output, best_logage_output, lmass_all, lage_all =\
----> 2     ysoisochrone.bayesian.derive_stellar_mass_age(df_prop, model='Baraffe_n_Feiden', plot=True)

File ~/anaconda3/envs/jossenv/lib/python3.11/site-packages/ysoisochrone/bayesian.py:348, in derive_stellar_mass_age(df_prop, model, isochrone_data_dir, isochrone_mat_file, no_uncertainties, plot, save_fig, save_lfunc, fig_save_dir, csv_save_dir, verbose, toofaint, toobright, median_age, confidence_interval, single_bayesian_for_nolum_target)
    345 lfunc[idx_age:, :] = 1e-99
    347 # Use Bayesian mass and age estimation
--> 348 bayes_res, lage_res, lmass_res = bayesian_mass_age(log_age_dummy, np.log10(masses_dummy), lfunc, plot=plot, source=source_t, verbose=verbose, save_fig=save_fig, fig_save_dir=fig_save_dir, confidence_interval=confidence_interval)
    349 best_logmass_wunc = [bayes_res[2], bayes_res[3][0], bayes_res[3][1]]
    350 best_logage_wunc = [bayes_res[0], bayes_res[1][0], bayes_res[1][1]]

File ~/anaconda3/envs/jossenv/lib/python3.11/site-packages/ysoisochrone/bayesian.py:57, in bayesian_mass_age(log_age_dummy, log_masses_dummy, L, plot, source, confidence_interval, verbose, save_fig, fig_save_dir, customized_fig_name)
     49 half_sigma_perc = confidence_interval/2.0
     51 # if verbose: # for debugging
     52 #     print('the shape of logage', log_age_dummy.shape)
     53 #     print('the shape of logmass', log_masses_dummy.shape)
     54 #     print('the shape of L', L.shape)
     55
     56 # Integrate L over mstar for each age bin
---> 57 L_age = np.array([simpson(L[jj, :], log_masses_dummy) for jj in range(len(log_age_dummy))])
     58 L_age /= simpson(L_age, log_age_dummy)
     60 # if verbose: # for debugging
     61 #     print('L_age:', L_age)
     62 #     print(np.cumsum(L_age))
     63 #     print(np.cumsum(L_age)/np.sum(L_age))

File ~/anaconda3/envs/jossenv/lib/python3.11/site-packages/ysoisochrone/bayesian.py:57, in <listcomp>(.0)
     49 half_sigma_perc = confidence_interval/2.0
     51 # if verbose: # for debugging
     52 #     print('the shape of logage', log_age_dummy.shape)
     53 #     print('the shape of logmass', log_masses_dummy.shape)
     54 #     print('the shape of L', L.shape)
     55
     56 # Integrate L over mstar for each age bin
---> 57 L_age = np.array([simpson(L[jj, :], log_masses_dummy) for jj in range(len(log_age_dummy))])
     58 L_age /= simpson(L_age, log_age_dummy)
     60 # if verbose: # for debugging
     61 #     print('L_age:', L_age)
     62 #     print(np.cumsum(L_age))
     63 #     print(np.cumsum(L_age)/np.sum(L_age))

TypeError: simpson() takes 1 positional argument but 2 were given
```
