### 2024-12-13 #7493 @HeloiseS
(A running list I'll update throughout my review of stuff that's not big enough to be put in an issue)

_Nice-to-have_
- [ ] Juypter (``notebook``) not listed as a dependency but needed to run the tutos
