### 2024-12-13 #7493 @HeloiseS
## Comments on the paper
- [ ]  YSOs should be expanded the first time (it's expanded the second time)
- [ ] line 35 and 36: looks like a list of bullet points didn't format properly
- [ ] line 44, the examples aren't showing up.

A more general question for @DingshanDeng: do none of the other track fitting algorithm allow you to use your own YSO tracks but with their fitting methods?

For example isn't there other open source codes where you could instead add the YSO functionality rather than create a whole new package?
- [ ] I'd like a bit more text on why a separate code is needed and joining other endeavours isn't workable.

For the paper that's all I have so you can get started on those changes as well.

-----

Note to @warrickball - I've reviewed paper and install. Started on tutorials but they are giving me an error so until [Issue #6](https://github.com/DingshanDeng/ysoisochrone/issues/6) is sorted I've done pretty much all I can do at this point.

If changes get made and I miss it (because GH floods my inbox with many things) feel free to poke me. This is now top of my list to get done and if I take more than 24h to show activity it's because I've missed something :)
