### 2024-12-14 DingshanDeng/ysoisochrone#6 @DingshanDeng
Hi thanks for reviewing the code and reporting this issue. What is the `SciPy` and `NumPy` versions are you using? It looks like the issue related to the `simpson` def from `SciPy`, and I will probably use an alternative one to do integral to fix it.
