### 2024-12-14 DingshanDeng/ysoisochrone#6 @warrickball
I haven't tried `ysoisochrone` at all but I suspect you've run into this [change in SciPy 1.12](https://docs.scipy.org/doc/scipy/release/1.12.0-notes.html#deprecated-features):

> Coinciding with upcoming changes to function signatures (e.g. removal of a deprecated keyword), we are deprecating positional use of keyword arguments for the affected functions, which will raise an error starting with SciPy 1.14. In some cases, this has delayed the originally announced removal date, to give time to respond to the second part of the deprecation. Affected functions are:
> * `linalg.{eigh, eigvalsh, pinv}`
> * `integrate.simpson`
> * `signal.{firls, firwin, firwin2, remez}`
> * `sparse.linalg.{bicg, bicgstab, cg, cgs, gcrotmk, gmres, lgmres, minres, qmr, tfqmr}`
> * `special.comb`
> * `stats.kendalltau`

So a line like
```python
L_age = np.array([simpson(L[jj, :], log_masses_dummy) for jj in range(len(log_age_dummy))])
```
probably needs to be something like
```python
L_age = np.array([simpson(L[jj, :], x=log_masses_dummy) for jj in range(len(log_age_dummy))])
```
but I repeat I haven't tried any code here. This should work in SciPy 1.12+ as well as older versions.
