### 2024-12-15 f733e75
attemp to fix the issue of calling scipy simpson
