### 2024-12-15 DingshanDeng/ysoisochrone#6 @DingshanDeng
Thank you Warrick! I think this is the issue. I updated the function following Warrick's suggestion, and tried to run the code with `SciPy v1.14`. It works now from my side. The updated version is now pushed to the main branch on the GitHub repo. Please give it a try to see whether it works for you now.
