### 2024-12-16 DingshanDeng/ysoisochrone#7 @HeloiseS
[JOSS] - Tutorial_2 -  ParserError: Error tokenizing data. C error: Expected 1 fields in line 7, saw 4
This is an issue relating to your submission to JOSS (see https://github.com/openjournals/joss-reviews/issues/7493).

Nearly everything ran I just encountered an issue on the last cell

When running

```
import pandas as pd

df_prop = pd.read_csv('example_targets.csv')

df_prop_t = df_prop.loc[:0]

best_logmass_output_t, best_logage_output_t, _, _ =\
    ysoisochrone.bayesian.derive_stellar_mass_age(df_prop_t, model='customize', plot=True,
                                                  isochrone_mat_file=output_mat_file, verbose=True)
```
I got:
```
ParserError                               Traceback (most recent call last)
Cell In[7], line 3
      1 import pandas as pd
----> 3 df_prop = pd.read_csv('example_targets.csv')
      5 df_prop_t = df_prop.loc[:0]
      7 best_logmass_output_t, best_logage_output_t, _, _ =\
      8     ysoisochrone.bayesian.derive_stellar_mass_age(df_prop_t, model='customize', plot=True,
      9                                                   isochrone_mat_file=output_mat_file, verbose=True)

File ~/anaconda3/lib/python3.10/site-packages/pandas/io/parsers/readers.py:1026, in read_csv(filepath_or_buffer, sep, delimiter, header, names, index_col, usecols, dtype, engine, converters, true_values, false_values, skipinitialspace, skiprows, skipfooter, nrows, na_values, keep_default_na, na_filter, verbose, skip_blank_lines, parse_dates, infer_datetime_format, keep_date_col, date_parser, date_format, dayfirst, cache_dates, iterator, chunksize, compression, thousands, decimal, lineterminator, quotechar, quoting, doublequote, escapechar, comment, encoding, encoding_errors, dialect, on_bad_lines, delim_whitespace, low_memory, memory_map, float_precision, storage_options, dtype_backend)
   1013 kwds_defaults = _refine_defaults_read(
   1014     dialect,
   1015     delimiter,
   (...)
   1022     dtype_backend=dtype_backend,
   1023 )
   1024 kwds.update(kwds_defaults)
-> 1026 return _read(filepath_or_buffer, kwds)

File ~/anaconda3/lib/python3.10/site-packages/pandas/io/parsers/readers.py:626, in _read(filepath_or_buffer, kwds)
    623     return parser
    625 with parser:
--> 626     return parser.read(nrows)

File ~/anaconda3/lib/python3.10/site-packages/pandas/io/parsers/readers.py:1923, in TextFileReader.read(self, nrows)
   1916 nrows = validate_integer("nrows", nrows)
   1917 try:
   1918     # error: "ParserBase" has no attribute "read"
   1919     (
   1920         index,
   1921         columns,
   1922         col_dict,
-> 1923     ) = self._engine.read(  # type: ignore[attr-defined]
   1924         nrows
   1925     )
   1926 except Exception:
   1927     self.close()

File ~/anaconda3/lib/python3.10/site-packages/pandas/io/parsers/c_parser_wrapper.py:234, in CParserWrapper.read(self, nrows)
    232 try:
    233     if self.low_memory:
--> 234         chunks = self._reader.read_low_memory(nrows)
    235         # destructive to chunks
    236         data = _concatenate_chunks(chunks)

File parsers.pyx:838, in pandas._libs.parsers.TextReader.read_low_memory()

File parsers.pyx:905, in pandas._libs.parsers.TextReader._read_rows()

File parsers.pyx:874, in pandas._libs.parsers.TextReader._tokenize_rows()

File parsers.pyx:891, in pandas._libs.parsers.TextReader._check_tokenize_status()

File parsers.pyx:2061, in pandas._libs.parsers.raise_parser_error()

ParserError: Error tokenizing data. C error: Expected 1 fields in line 7, saw 4

```
