### 2024-12-20 dfda8bf
update the ploting code and the tutorial
