### 2024-12-20 DingshanDeng/ysoisochrone#7 @DingshanDeng
Hi, thanks for reporting this issue. The error is because there are some comment lines starting with '#' in the example csv file. Now I made some change to tutorial 2, by setting up a new example target with in the script directly instead of reading in a new file. I hope this would be a clearer way to demonstrate in the tutorial. The new version is uploaded to github/main branch.
