### 2024-12-28 DingshanDeng/ysoisochrone#2 @vsquicciarini
Hi @DingshanDeng, when installing the package I had the same thought that @HeloiseS explained well above.

Regarding point 2), on my side I strongly suggest using the Python Package Index (PyPI) to make the package installable via pip. This is especially useful because it gives you full control of the versions of your package you want people to use and because it allows for an easier interaction with, e.g., virtual environments. I may provide help with that, if you want to.
