### 2024-12-28 #7493 @vsquicciarini
Hi @DingshanDeng, first of all I would like to apologize with you for the long wait. Being amidst a job change with a spare month between the two contracts, I thought I could have had much more free time than I eventually had. I am truly sorry about that.

I have read the paper, the documentation and tested the Jupyter Notebook you provided us with. My overall impression of your work is positive, even though I think that are a few issues to be discussed together. I am going to write down comments related to the paper and the documentation. I would like to keep on testing the code on my own for a couple of additional days before discussing the package from a technical point of view.
