### 2024-12-28 #7493 @vsquicciarini
## Comments on the paper

- [ ] line 9: stellar age -> stellar ages.
- [ ] line 15: I would use the singular: the stellar mass, age, its effective temperature etc.
- [ ] line 18: please add references here for "as assumed in previous literature". The point here is not trivial: to me, a log-uniform prior for Teff and L is more physically sound than a uniform prior. The assumption of linearly uniform priors is not uninformative, because it creates a bias for hotter and brighter stars. Both the shape of the initial mass function of stars and their evolutionary timescales imply that the occurrence of stars decreases as a function of Teff and L. Hence, using a log-uniform prior would give a more unbiased estimates of the underlying distribution. I know that this is a strong request, but I would suggest replacing the observables Teff and L with logTeff and logL.
- [ ] line 24: it should be specified that the Feiden models here are non-magnetic. I could find this information from the quoted Pascucci+16 reference, but I think that this should be explicitly stated.
- [ ] line 24: the choice of combining Baraffe and Feiden models is based on the study by Herczeg & Hillenbrand (2015), quoted in Pascucci+16. While I do not question their models and results, I note that over the last 10 years there were both significant improvements of stellar models and the observational revolution brought about by Gaia. My point is that a more recent study should be used here to justify the choice of the isochrone set. If this is not possible, I think that it would be perfectly fine to state more explicitly that your choice is based on the 2015 analysis, that might not be up to date, and that you intend to add new models in later versions of your package.
- [ ] general comment: I would appreciate if you could detail in the text why you chose to use as input parameters L and Teff, instead of the measured photometry. On my side, I find this choice problematic for several reasons (that I will detail below); but I am not really an expert of proto-stars and <5 Myr-old stars with a full disk, so there might be reasons for your choice that I am overlooking.
- [ ] statement of need: I think this has to be better rephrased: the e.g. parethesis is currently empty, and it is not clear what the advantage of this package over previous work is. I'm sorry to say that, but I have written a fully documented and open source package, MADYS (https://madys.readthedocs.io/en/latest/), that can be used to derive ages and masses for pre-main sequence stars. Another package, isochrones, is cited in the documentation but not here. I think that both packages should be, if not discussed, at least acknowledged here.

### On the choice of input observables

As I mentioned before, I think it is necessary to explain why (L, Teff) were chosen over photometry as input variables. Based on my experience, I consider the latter to be much more reliable than the former because:
1) luminosity and effective temperatures are not observed quantities, but they are derived from photometry assuming some atmospheric model. Photometry is a real observable, only dependent on the filter response and the zero point of the instrument;
2) being model-dependent, two measurements of L and Teff from two different papers might assume different underlying models, ages, parallaxes, metallicities, bolometric corrections, and be based on different observational data (photometry, spectroscopy); this makes any direct comparison of input data extremely difficult;
3) as a consequence of 1) and 2), a (L, Teff) point obtained through a model A and then fitted by your package to model B will be affected by a systematic error which is not just a random error, but also comprises the systematic difference between model A and model B. How is this taken into account in the error estimation?
