### 2024-12-28 #7493 @vsquicciarini
## Some comments on the Jupyter Notebook

As I mentioned above, I may write some additional technical comments on the code itself after testing it more carefully. For the moment being, I just write here some general comments that came to my mind while executing the notebook:

- [ ] 4.3) Estimate mass and ages: I understand the practical advantage of attempting to estimate masses even for underluminous stars. However, I think that the results here should be taken with extreme caution, and the user should be aware of that. I would add a flag column in `df_output` that specifies if the star is within the grid, underluminous or overluminous. Also, I notice that no error is provided on the mass and age here. This should be fixed, because 1) uncertainties must always be estimated, 2) for non-standard stars is it even more crucial to understand how much the mass is constrained. The uncertainty on Teff and the uncertainty on the assume age should be used to evaluate the corresponding mass uncertainty.
- [ ] 5.1) Closest grid point on isochrone: as above, I think it is mandatory to provide uncertainty estimation here. I would estimate the uncertainties in a Monte Carlo way, i.e. by picking the closest (mass, age) grid point but with L and Teff randomly altered N times (let's say, 1000) according to their uncertainties. While I understand that this is not the preferred method and it is only provided to reproduce literature results, I stress the importance of doing that to allow for meaningful comparison with the literature itself.
- [ ] 5.2) Assuming age to derive stellar masses: it is not clear to me what the difference between the two possibilities is. Am I right in thinking that in the first case it interpolates across mass, while in the second just takes the closest track? I think this should be better rephrased.
- [ ] general comment: if a star is underluminous due to disk absorption, we should expect an effect on temperature as well. This means that when we use an age and a Teff, the Teff should be biased to lower values too (due to differential absorption). Shouldn't this be taken into account by using some kind of de-reddening? Also, depending on the way Teff was derived, Teff itself might be more or less biased. For instance, if Teff was derived only using NIR data, then the disk luminosity would introduce a larger bias compared to cases where UV measurements are available. Again, this stresses how the choice of Teff as an input parameter can induce a lot of unseen systematics.
