### 2024-12-28 #7493 @vsquicciarini
## Comments on the documentation

- [ ] "The available stellar evolutionary models": in the sentence "This package is mainly built to study the young-stellar-objects, so by default; it only utilizes the stellar evolutionary tracks for pre-main-sequence stars ({\rm stellar\ age}< 500\,{\rm Myrs}). " , I think the parenthesis is misleading because the duration of the pre-MS strongly depends on mass. A solar mass star only stays on the pre-MS for about 50 Myr. I suggest removing the parenthesis.
- [ ] In general, I find that the paragraph above could be rewritten better. I suggest: "For the moment being, this package is meant to be used to study young stellar objects. In other words, it only utilizes the stellar evolutionary tracks for pre-main-sequence stars, avoiding the problem of dealing with main-sequence and post-main-sequence targets (whose luminosity rises again and will overlay on the pre-main-sequence phase). In future releases, the package might be expanded to also handle main-sequence stars (e.g., Fernandes et al. 2023), but user-specified customized stellar evolutionary tracks would be needed."
- [ ] Feiden: again, it should be specified that the isochrones are non-magnetic.
- [ ] MIST: since you are quoting the `isochrones` package, I suggest quoting my `MADYS` package as well, which can handle MIST, PARSEC (1.2 and 2.0), Feiden, Baraffe and many other models for pre-MS or MS stars.
