### 2025-01-13 dea3e0d
update readme with comminuty guidelines; update paper; preparing pytest files
