+++ paper/paper.bib
@@ -108,6 +108,26 @@ archivePrefix = {arXiv},
 }
 
 
+@ARTICLE{herczeg_empirical_2015,
+	title = {Empirical {Isochrones} for {Low} {Mass} {Stars} in {Nearby} {Young} {Associations}},
+	volume = {808},
+	issn = {0004-637X},
+	url = {https://ui.adsabs.harvard.edu/abs/2015ApJ...808...23H},
+	doi = {10.1088/0004-637X/808/1/23},
+	abstract = {Absolute ages of young stars are important for many issues in pre-main-sequence stellar and circumstellar evolution but long have been recognized as difficult to derive and calibrate. In this paper, we use literature spectral types and photometry to construct empirical isochrones in Hertzsprung-Russell diagrams for low mass stars and brown dwarfs in the η Cha, ɛ Cha, and TW Hya Associations and the β Pic and Tuc-Hor Moving Groups. A successful theory of pre-main-sequence evolution should match the shapes of the stellar loci for these groups of young stars. However, when comparing the combined empirical isochrones to isochrones predicted from evolutionary models, discrepancies lead to a spectral type (mass) dependence in stellar age estimates. Improved prescriptions for convection and boundary conditions in the latest models of pre-main-sequence evolution lead to a significantly improved correspondence between empirical and model isochrones, with small offsets at low temperatures that may be explained by observational uncertainties or by model limitations. Independent of model predictions, linear fits to combined stellar loci of these regions provide a simple empirical method to order clusters by luminosity with a reduced dependence on spectral type. Age estimates calculated from various sets of modern models that reproduce Li depletion boundary ages of the β Pic Moving Group also imply a ∼4 Myr age for the low mass members of the Upper Sco OB Association, which is younger than the 11 Myr age that has been recently estimated for intermediate and high mass members.},
+	urldate = {2024-09-28},
+	journal = {The Astrophysical Journal},
+	author = {Herczeg, Gregory J. and Hillenbrand, Lynne A.},
+	month = jul,
+	year = {2015},
+	note = {Publisher: IOP
+ADS Bibcode: 2015ApJ...808...23H},
+	keywords = {Astrophysics - Solar and Stellar Astrophysics, stars: evolution, stars: formation, Hertzsprung–Russell and C–M diagrams},
+	pages = {23},
+	file = {Accepted Version:/Users/dingshandeng/Zotero/storage/BNZXGD3G/Herczeg and Hillenbrand - 2015 - Empirical Isochrones for Low Mass Stars in Nearby Young Associations.pdf:application/pdf},
+}
+
+
 @ARTICLE{Pascucci_2016,
        author = {{Pascucci}, I. and {Testi}, L. and {Herczeg}, G.~J. and {Long}, F. and {Manara}, C.~F. and {Hendler}, N. and {Mulders}, G.~D. and {Krijt}, S. and {Ciesla}, F. and {Henning}, Th. and {Mohanty}, S. and {Drabek-Maunder}, E. and {Apai}, D. and {Sz{\H{u}}cs}, L. and {Sacco}, G. and {Olofsson}, J.},
         title = "{A Steeper than Linear Disk Mass-Stellar Mass Scaling Relation}",
