+++ paper/paper.md
@@ -27,13 +27,13 @@ bibliography: paper.bib
 ---
 
 # Background and Methods
-There has been a long history of estimating stellar age and masses from the stellar evolutionary models (e.g., @Siess_2000, @Feiden_2016, @Baraffe_2015). Different methods have been employed, from finding the closest track to an object's luminosity and temperature (e.g., @Manara_2023_PPVII) to employing a Bayesian approach which enables estimating uncertainties on the inferred ages and masses (e.g., @Jorgensen_n_Lindergren_2005, @Gennaro_2012, @Andrews_2013). Our primary method is a Bayesian inference approach, and the `Python` code builds on the `IDL` version developed in @Pascucci_2016. The code estimates the stellar masses, ages, and associated uncertainties by comparing their stellar effective temperature ($T_{\rm eff}$), bolometric luminosity ($L_{\rm bol}$), and their uncertainties with different stellar evolutionary models, including those specifically developed for YSOs. As assumed in previous literature, we adopt uniforms priors, and the likelihood functions are defined following:
+There has been a long history of estimating stellar age and masses from the stellar evolutionary models (e.g., @Siess_2000, @Feiden_2016, @Baraffe_2015). Different methods have been employed, from finding the closest track to an object's luminosity and temperature (e.g., @Manara_2023_PPVII) to employing a Bayesian approach which enables estimating uncertainties on the inferred ages and masses (e.g., @Jorgensen_n_Lindergren_2005, @Gennaro_2012, @Andrews_2013). Our primary method is a Bayesian inference approach, and the `Python` code builds on the `IDL` version developed in @Pascucci_2016. The code estimates the stellar masses, ages, and associated uncertainties by comparing their stellar effective temperature ($T_{\rm eff}$), bolometric luminosity ($L_{\rm bol}$), and their uncertainties with different stellar evolutionary models, including those specifically developed for young stellar objects (YSOs). As assumed in previous literature, we adopt uniforms priors, and the likelihood functions are defined following:
 $$
 \mathcal{L}(T_{i}, L_{i} | T_{\rm obs}, L_{\rm obs}) = \frac{1}{2\pi \sigma_{T_{\rm obs}} \sigma_{L_{\rm obs}}} \exp\left( -\frac{1}{2} \left[ \frac{(T_{\rm obs} - T_{i})^2}{\sigma_{T_{\rm obs}}^2} + \frac{(L_{\rm obs} - L_{i})^2}{\sigma_{L_{\rm obs}}^2} \right] \right), \tag{1}
 $$
 where the $T$ and $L$ are the effective temperature and bolometric luminosity, respectively. $T_i, L_i$ are the values from the evolutionary model grids, and $T_{\rm obs}, L_{\rm obs}$ are the observed values for each target with their uncertainties ($1\,\sigma$ from the assumed Gaussian distribution) described as $\sigma_{T_{\rm obs}}, \sigma_{L_{\rm obs}}$.
 
[…]
