+++ requirements.txt
@@ -4,3 +4,4 @@ matplotlib
 scipy
 requests
 tqdm
+jupyter
