+++ README.md
@@ -37,3 +37,6 @@ This Guide is also provided in the [documentation](https://ysoisochrone.readthed
 
 ## Citations
 If you use `ysoisochrone` as part of your research, please cite the xxx
+
+## Community Guidelines
+We welcome contributions, issue reports, and questions about `ysoisochrone`! If you encounter a bug or issue, check out the [Issues page](https://github.com/DingshanDeng/ysoisochrone/issues) and provide a report with details about the problem and steps to reproduce it. For general support, usage questions and suggestions, you can start a discussion in [Discussions page](https://github.com/DingshanDeng/ysoisochrone/discussions), and of course feel free to send emails directly to us. If you want to contribute, feel free to fork the repository and create pull requests here. `ysoisochrone` is licensed under MIT license, so feel free to make use of the source code in any part of your own work/software.
