### 2025-01-14 d402969
adding unit test using pytest; appending this to github workflow
