### 2025-01-14 DingshanDeng/ysoisochrone#4 @DingshanDeng
Hi, the comminuty guideline is now added to the readme.md in the home page. Hope that is clear.
