+++ .github/workflows/pytesting.yml
@@ -26,6 +26,7 @@ jobs:
           python -m pip install --upgrade pip
           pip install -r requirements.txt
           pip install pytest
+          pip install -e .  # Install the current package in editable mode
 
       # Run pytest
       - name: Run tests
