### 2025-01-14 DingshanDeng/ysoisochrone#3 @DingshanDeng
Hi, thanks for pointing this out and also thank for your helpful suggestions. I implemented `pytest` to do the unit test also following your suggestion on using the github workflow to run the test. It looks like it went through successfully.
