### 2025-01-15 DingshanDeng/ysoisochrone#2 @DingshanDeng
Hi, the package has been uploaded to PyPI (https://pypi.org/project/ysoisochrone/0.4.0b0) and should be able to be installed via `pip install ysoisochrone`. Thanks for your suggestions!
