### 2025-01-15 f48825e
uploaded the package to PyPI, and working on the new doc and paper
