+++ README.md
@@ -21,7 +21,13 @@ Rachel B. Fernandes, The Pennsylvania State University
 
 ## Installation
 
-First download the package release from the GitHub page. Then unzip the package.
+You can easily install the package via
+
+```bash
+pip install ysoisochrone
+```
+
+Or, you can also install your preferred release by downloading the package release from the GitHub page. Then unzip the package.
 
 In the terminal and in the directory of this package where `setup.py` exists.
 
