### 2025-01-16 DingshanDeng/ysoisochrone#2 @HeloiseS
I created a conda environment form scratch and did the pip install.
Both notebooks run fully so I am happy to close this issue when @vsquicciarini is happy too.
