### 2025-01-16 83246e1
update paper and documentation
