+++ paper/paper.bib
@@ -1,3 +1,23 @@
+@article{squicciarini_madys_2022,
+	title = {{MADYS}: the {Manifold} {Age} {Determination} for {Young} {Stars}. {I}. {Isochronal} age estimates and model comparison},
+	volume = {666},
+	issn = {0004-6361},
+	shorttitle = {{MADYS}},
+	url = {https://ui.adsabs.harvard.edu/abs/2022A&A...666A..15S},
+	doi = {10.1051/0004-6361/202244193},
+	abstract = {Context. The unrivalled astrometric and photometric capabilities of the Gaia mission have given new impetus to the study of young stars: both from an environmental perspective, as members of comoving star-forming regions, and from an individual perspective, as targets amenable to planet-hunting direct-imaging observations. Aims: In view of the large availability of theoretical evolutionary models, both fields would benefit from a unified framework that allows a straightforward comparison of physical parameters obtained by different stellar and substellar models. Methods: To this aim, we developed the Manifold Age Determination for Young Stars (madys), a flexible Python tool for the age and mass determination of young stellar and substellar objects. In this first release, madys automatically retrieves and crossmatches photometry from several catalogs, estimates interstellar extinction, and derives age and mass estimates for individual objects through isochronal fitting. Results: Harmonizing the heterogeneity of publicly available isochrone grids, the tool allows one to choose amongst 17 models, many of which with customizable astrophysical parameters, for a total of {\textasciitilde}110 isochrone grids. Several dedicated plotting functions are provided to allow for an intuitive visual perception of the numerical output. Conclusions: After extensive testing, we have made the tool publicly available. Here, we demonstrate the capabilities of madys, summarizing previously published results as well providing several new examples.},
+	urldate = {2025-01-16},
+	journal = {Astronomy and Astrophysics},
+	author = {Squicciarini, V. and Bonavita, M.},
+	month = oct,
+	year = {2022},
+	note = {ADS Bibcode: 2022A\&A...666A..15S},
+	keywords = {Astrophysics - Astrophysics of Galaxies, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics, methods: data analysis, planets and satellites: fundamental parameters, stars: fundamental parameters},
+	pages = {A15},
+	file = {Full Text PDF:/Users/dingshandeng/Zotero/storage/84GB3YZC/Squicciarini and Bonavita - 2022 - MADYS the Manifold Age Determination for Young Stars. I. Isochronal age estimates and model compari.pdf:application/pdf},
+}
+
+
 @INPROCEEDINGS{Soderblom_2014_PPVI,
        author = {{Soderblom}, D.~R. and {Hillenbrand}, L.~A. and {Jeffries}, R.~D. and {Mamajek}, E.~E. and {Naylor}, T.},
         title = "{Ages of Young Stars}",
@@ -50,6 +70,26 @@ archivePrefix = {arXiv},
  primaryClass = {astro-ph},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2000A&A...358..593S},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@article{simon_2019,
+	title = {Masses and {Implications} for {Ages} of {Low}-mass {Pre}-main-sequence {Stars} in {Taurus} and {Ophiuchus}},
+	volume = {884},
+	issn = {0004-637X},
+	url = {https://ui.adsabs.harvard.edu/abs/2019ApJ...884...42S},
+	doi = {10.3847/1538-4357/ab3e3b},
[…]
