+++ docs/source/index.md
@@ -15,7 +15,13 @@ We also provide two other ways to estimate the stellar masses and ages from thes
 
 ## Installation
 
-In the terminal and in the directory of this package where `setup.py` exists.
+You can easily install the package via
+
+```bash
+pip install ysoisochrone
+```
+
+Or, you can also install your preferred release by downloading the package release from the GitHub page. Then unzip the package. In the terminal and in the directory of this package where `setup.py` exists.
 
 ```bash 
 pip install .
@@ -38,6 +44,16 @@ Then you can start check out the [Quick Start Guide](./notebooks/ysoisochrone_ba
 If you end up using this package, please cite xxx (TBA)
 
 If you use any [stellar evolutionary models](./models.md), please also refer to their original work/website for citations.
+
+## Useful links
+
+There are a few other useful tools and packages that can be used to handle stellar evolutionary tracks and to estimate stellar mass and age for pre-main sequence stars. Including:
+
+- [`MADYS`](https://madys.readthedocs.io/en/latest/) is `Python` package that can be used to derive ages and masses for pre-main sequence stars from multi-wavelengths photometric data with the extinction corrected and could ustilize different stellar evolutionary models, including MIST, PARSEC (v1.2 and 2.0), Feiden, Baraffe and many other models for pre-MS or MS stars.
+
+- PARSEC team provides a [web interface](http://stev.oapd.inaf.it/PARSEC/tools.html) to access different versions of their tracks together with some useful web-based tools. 
+  
+- [`isochrones`](https://github.com/timothydmorton/isochrones) is a `Python` package that provides interface to access the [MIST](https://waps.cfa.harvard.edu/MIST/) grids.
 
 ## Contents:
 ```{toctree}

[…]
