### 2025-01-20 f567432
update paper. working on doc and notebook as well as adding flag
