+++ paper/paper.bib
@@ -262,3 +262,19 @@ archivePrefix = {arXiv},
 }
 
 
+@article{alcala_x-shooter_2017,
+	title = {X-shooter spectroscopy of young stellar objects in {Lupus}: {Accretion} properties of class {II} and transitional objects⋆},
+	volume = {600},
+	issn = {0004-6361, 1432-0746},
+	shorttitle = {X-shooter spectroscopy of young stellar objects in {Lupus}},
+	url = {http://www.aanda.org/10.1051/0004-6361/201629929},
+	doi = {10.1051/0004-6361/201629929},
+	language = {en},
+	urldate = {2024-03-14},
+	journal = {Astronomy \& Astrophysics},
+	author = {Alcalá, J. M. and Manara, C. F. and Natta, A. and Frasca, A. and Testi, L. and Nisini, B. and Stelzer, B. and Williams, J. P. and Antoniucci, S. and Biazzo, K. and Covino, E. and Esposito, M. and Getman, F. and Rigliaco, E.},
+	month = apr,
+	year = {2017},
+	pages = {A20},
+	file = {Alcalá et al. - 2017 - X-shooter spectroscopy of young stellar objects in.pdf:/Users/dingshandeng/Zotero/storage/KXHX9I8U/Alcalá et al. - 2017 - X-shooter spectroscopy of young stellar objects in.pdf:application/pdf},
+}
