+++ ysoisochrone/plotting.py
@@ -469,24 +469,27 @@ def plot_likelihood_1d(log_masses_dummy, likelihood, best_log_mass, lower_mass, 
     
     likelihood[likelihood <= 1e-98] = np.nan
     
-    plt.figure(figsize=(8, 6))
-    plt.plot(10**log_masses_dummy, likelihood, label='Likelihood', lw=2)
-    plt.axvline(x=10**best_log_mass, color='r', linestyle='--', label=f'Best Mass: {10**best_log_mass:.2f} $M_\\odot$')
-    plt.axvline(x=10**lower_mass, color='k', linestyle=':', label=f'Lower Bound: {10**lower_mass:.2f} $M_\\odot$')
-    plt.axvline(x=10**upper_mass, color='k', linestyle=':', label=f'Upper Bound: {10**upper_mass:.2f} $M_\\odot$')
+    fig = plt.figure(figsize=(8, 6))
+    ax  = fig.add_subplot(1,1,1)
+    ax.plot(10**log_masses_dummy, likelihood, label='Likelihood', lw=2)
+    ax.axvline(x=10**best_log_mass, color='r', linestyle='--', label=f'Best Mass: {10**best_log_mass:.2f} $M_\\odot$')
+    ax.axvline(x=10**lower_mass, color='k', linestyle=':', label=f'Lower Bound: {10**lower_mass:.2f} $M_\\odot$')
+    ax.axvline(x=10**upper_mass, color='k', linestyle=':', label=f'Upper Bound: {10**upper_mass:.2f} $M_\\odot$')
 
-    plt.xlabel(r'Stellar Mass [$M_\odot$]', fontsize=14)
-    plt.ylabel('Likelihood', fontsize=14)
-    # plt.xscale('log')
-    plt.yscale('log')
+    ax.set_xlabel(r'Stellar Mass [$M_\odot$]', fontsize=14)
+    ax.set_ylabel('Likelihood', fontsize=14)
+    # ax.set_xscale('log')
+    ax.set_yscale('log')
+    ylim_t = ax.get_ylim()
+    ax.set_ylim(bottom=np.nanmax(ylim_t)*1e-15)
     
     if source:
-        plt.title(f'Likelihood for Stellar Mass ({source})', fontsize=16)
+        ax.set_title(f'Likelihood for Stellar Mass ({source})', fontsize=16)
     else:
-        plt.title('Likelihood for Stellar Mass', fontsize=16)
+        ax.set_title('Likelihood for Stellar Mass', fontsize=16)
 
-    plt.legend(loc='upper left')
-    plt.grid(True)
+    ax.legend(loc='lower center')
+    ax.grid(True)
     plt.show()
     
     return 1
@@ -522,56 +525,56 @@ def plot_comparison(log_age_idl, masses_idl, logtlogl_interp_py, logtlogl_idl, l
     # Teff Plot (Interpolated Python Grid)
     im1 = axs[0, 0].imshow(logtlogl_interp_py[:, :, 0], aspect='auto', extent=extent)
     axs[0, 0].set_title('%s Grid (Interpolated to IDL): Teff'%(gridnames[0]))
-    axs[0, 0].set_xlabel('Mass [M$_\odot$]')
+    axs[0, 0].set_xlabel(r'Mass [M$_\odot$]')
     axs[0, 0].set_ylabel('log(Age) [years]')
     fig.colorbar(im1, ax=axs[0, 0])
 
     # Teff Plot (IDL Grid)
[…]
