### 2025-01-20 49295be
update the documentation and tutorial notebook; add the flags to output
