+++ paper/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 ---
 
 # Background and Methods
-There has been a long history of estimating **stellar ages and masses** from the stellar evolutionary models [e.g., @Siess_2000; @Feiden_2016; @Baraffe_2015]. Different methods have been employed, from finding the closest track to an object's luminosity and temperature [e.g., @Manara_2023_PPVII] to employing a Bayesian approach which enables estimating uncertainties on the inferred ages and masses [e.g., @Jorgensen_n_Lindergren_2005; @Gennaro_2012; @Andrews_2013]. Our primary method is a Bayesian inference approach, and the `Python` code builds on the `IDL` version developed in @Pascucci_2016. **The code estimates the stellar mass, age, and associated uncertainties by comparing its stellar effective temperature ($T_{\rm eff}$), bolometric luminosity ($L_{\rm bol}$)**, and their uncertainties with different stellar evolutionary models, including those specifically developed for young stellar objects **(YSOs)**. **Follow the same approach in the literature (e.g., @Andrews_2013, @Pascucci_2016), we compute a conditional likelihood function assuming uniform priors on the model parameters (since we have no prior knowledge to most targets):**
+There has been a long history of estimating **stellar ages and masses** from the stellar evolutionary models [@Siess_2000, prefix="e.g., "; @Feiden_2016; @Baraffe_2015]. Different methods have been employed, from finding the closest track to an object's luminosity and temperature [e.g., @Manara_2023_PPVII] to employing a Bayesian approach which enables estimating uncertainties on the inferred ages and masses (e.g., [@Jorgensen_n_Lindergren_2005]; [@Gennaro_2012]; [@Andrews_2013]). Our primary method is a Bayesian inference approach, and the `Python` code builds on the `IDL` version developed in @Pascucci_2016. **The code estimates the stellar mass, age, and associated uncertainties by comparing its stellar effective temperature ($T_{\rm eff}$), bolometric luminosity ($L_{\rm bol}$)**, and their uncertainties with different stellar evolutionary models, including those specifically developed for young stellar objects **(YSOs)**. **Follow the same approach in the literature (e.g., @Andrews_2013, @Pascucci_2016), we compute a conditional likelihood function assuming uniform priors on the model parameters (since we have no prior knowledge to most targets):**
 $$
 \mathcal{L}(T_{i}, L_{i} | T_{\rm obs}, L_{\rm obs}) = \frac{1}{2\pi \sigma_{T_{\rm obs}} \sigma_{L_{\rm obs}}} \exp\left( -\frac{1}{2} \left[ \frac{(T_{\rm obs} - T_{i})^2}{\sigma_{T_{\rm obs}}^2} + \frac{(L_{\rm obs} - L_{i})^2}{\sigma_{L_{\rm obs}}^2} \right] \right), \tag{1}
 $$
