+++ paper/paper.md
@@ -27,14 +27,15 @@ bibliography: paper.bib
 ---
 
 # Background and Methods
-There has been a long history of estimating **stellar ages and masses** from the stellar evolutionary models [e.g., @Baraffe_2015; @Feiden_2016; @Siess_2000]. Different methods have been employed, from finding the closest track to an object's luminosity and temperature [e.g., @Manara_2023_PPVII] to employing a Bayesian approach which enables estimating uncertainties on the inferred ages and masses [e.g., @Andrews_2013; @Gennaro_2012;  @Jorgensen_n_Lindergren_2005]. Our primary method is a Bayesian inference approach, and the `Python` code builds on the `IDL` version developed in @Pascucci_2016. **The code estimates the stellar mass, age, and associated uncertainties by comparing its stellar effective temperature ($T_{\rm eff}$), bolometric luminosity ($L_{\rm bol}$)**, and their uncertainties with different stellar evolutionary models, including those specifically developed for young stellar objects **(YSOs)**. **Follow the same approach in the literature [e.g., @Andrews_2013; @Pascucci_2016], we compute a conditional likelihood function assuming uniform priors on the model parameters (since we have no prior knowledge to most targets):**
+There has been a long history of estimating **stellar ages and masses** from the stellar evolutionary models [e.g., @Baraffe_2015; @Feiden_2016; @Siess_2000]. Different methods have been employed, from finding the closest track to an object's luminosity and temperature [e.g., @Manara_2023_PPVII] to employing a Bayesian approach which enables estimating uncertainties on the inferred ages and masses [e.g., @Andrews_2013; @Gennaro_2012;  @Jorgensen_n_Lindergren_2005]. Our primary method is a Bayesian inference approach, and the `Python` code builds on the `IDL` version developed by @Pascucci_2016. **The code estimates the stellar mass, age, and associated uncertainties by comparing a star's effective temperature ($T_{\rm eff}$), bolometric luminosity ($L_{\rm bol}$)**, and their uncertainties with different stellar evolutionary models, including those specifically developed for young stellar objects **(YSOs)**. **The conditional likelihood function assumes uniform priors and can be written as:**
 $$
 \mathcal{L}(T_{i}, L_{i} | T_{\rm obs}, L_{\rm obs}) = \frac{1}{2\pi \sigma_{T_{\rm obs}} \sigma_{L_{\rm obs}}} \exp\left( -\frac{1}{2} \left[ \frac{(T_{\rm obs} - T_{i})^2}{\sigma_{T_{\rm obs}}^2} + \frac{(L_{\rm obs} - L_{i})^2}{\sigma_{L_{\rm obs}}^2} \right] \right), \tag{1}
 $$
 where the $T$ and $L$ are the effective temperature and bolometric luminosity, respectively. $T_i, L_i$ are the values from the evolutionary model grids, and $T_{\rm obs}, L_{\rm obs}$ are the observed values for each target with their uncertainties ($1\,\sigma$ from the assumed Gaussian distribution) described as $\sigma_{T_{\rm obs}}, \sigma_{L_{\rm obs}}$.
-**We choose $T_{\rm eff}$ and $L_{\rm bol}$ to estimate the stellar age and mass because the extinction is significant for young stars especially those embeded objects. Although the $T_{\rm eff}$ and $L_{\rm bol}$ are not directly observed quantities, they are the two main quantities that evolutionary models can be compared with. When median or high-resolution spectroscopy is employed on individual targets, the $T_{\rm eff}$ and $L_{\rm bol}$ can be better determined, and the best estimates for YSOs are from works where a stellar spectrum is fitted simutaneously with extinction and accretional heating [e.g., @alcala_x-shooter_2017].**
[…]
