### 2025-01-22 f0cb20f
update paper document and tutorials
