+++ paper/paper.md
@@ -22,20 +22,21 @@ affiliations:
    index: 3
  - name: Center for Exoplanets and Habitable Worlds, 525 Davey Laboratory, The Pennsylvania State University, University Park, PA 16802, USA
    index: 4
-date: 21 Janurary 2025
+date: 22 Janurary 2025
 bibliography: paper.bib
 ---
 
 # Background and Methods
-There has been a long history of estimating **stellar ages and masses** from the stellar evolutionary models [e.g., @Baraffe_2015; @Feiden_2016; @Siess_2000]. Different methods have been employed, from finding the closest track to an object's luminosity and temperature [e.g., @Manara_2023_PPVII] to employing a Bayesian approach which enables estimating uncertainties on the inferred ages and masses [e.g., @Andrews_2013; @Gennaro_2012;  @Jorgensen_n_Lindergren_2005]. Our primary method is a Bayesian inference approach, and the `Python` code builds on the `IDL` version developed by @Pascucci_2016. **The code estimates the stellar mass, age, and associated uncertainties by comparing a star's effective temperature ($T_{\rm eff}$), bolometric luminosity ($L_{\rm bol}$)**, and their uncertainties with different stellar evolutionary models, including those specifically developed for young stellar objects **(YSOs)**. **The conditional likelihood function assumes uniform priors and can be written as:**
+There has been a long history of estimating **stellar ages and masses** from the stellar evolutionary models [e.g., @Baraffe_2015; @Feiden_2016; @Siess_2000]. Different methods have been employed, from finding the closest track to an object's luminosity and temperature [e.g., @Manara_2023_PPVII] to employing a Bayesian approach which enables estimating uncertainties on the inferred ages and masses [e.g., @Andrews_2013; @Gennaro_2012;  @Jorgensen_n_Lindergren_2005]. Our primary method is a Bayesian inference approach, and the `Python` code builds on the `IDL` version developed by @Pascucci_2016. **The code estimates the stellar mass, age, and associated uncertainties by comparing a star's effective temperature ($T_{\rm eff}$), bolometric luminosity ($L_{\rm bol}$)**, and their uncertainties with different stellar evolutionary models, including those specifically developed for young stellar objects **(YSOs)**. **The conditional likelihood function assumes log-uniform priors and can be written as:**
 $$
-\mathcal{L}(T_{i}, L_{i} | T_{\rm obs}, L_{\rm obs}) = \frac{1}{2\pi \sigma_{T_{\rm obs}} \sigma_{L_{\rm obs}}} \exp\left( -\frac{1}{2} \left[ \frac{(T_{\rm obs} - T_{i})^2}{\sigma_{T_{\rm obs}}^2} + \frac{(L_{\rm obs} - L_{i})^2}{\sigma_{L_{\rm obs}}^2} \right] \right), \tag{1}
+\mathcal{L}(T_{i}, L_{i} | T_{\rm obs}, L_{\rm obs}) = \frac{1}{2\pi \sigma_{\log_{10}T_{\rm obs}} \sigma_{\log_{10}L_{\rm obs}}} \exp\left( -\frac{1}{2} \left[ \frac{(\log_{10}T_{\rm obs} - \log_{10}T_{i})^2}{\sigma_{\log_{10}T_{\rm obs}}^2} + \frac{(\log_{10}L_{\rm obs} - \log_{10}L_{i})^2}{\sigma_{\log_{10}L_{\rm obs}}^2} \right] \right), \tag{1}
 $$
[…]
