+++ paper/paper.md
@@ -28,11 +28,21 @@ bibliography: paper.bib
 
 # Background and Methods
 There has been a long history of estimating **stellar ages and masses** from the stellar evolutionary models [e.g., @Baraffe_2015; @Feiden_2016; @Siess_2000]. Different methods have been employed, from finding the closest track to an object's luminosity and temperature [e.g., @Manara_2023_PPVII] to employing a Bayesian approach which enables estimating uncertainties on the inferred ages and masses [e.g., @Andrews_2013; @Gennaro_2012;  @Jorgensen_n_Lindergren_2005]. Our primary method is a Bayesian inference approach, and the `Python` code builds on the `IDL` version developed by @Pascucci_2016. **The code estimates the stellar mass, age, and associated uncertainties by comparing a star's effective temperature ($T_{\rm eff}$), bolometric luminosity ($L_{\rm bol}$)**, and their uncertainties with different stellar evolutionary models, including those specifically developed for young stellar objects **(YSOs)**. **The conditional likelihood function assumes log-uniform priors and can be written as:**
+<!-- $$
+\mathcal{L}(\log_{10}T_{i}, \log_{10}L_{i} | \log_{10}T_{\rm obs}, \log_{10}L_{\rm obs}) = \frac{1}{2\pi \sigma_{\log_{10}T_{\rm obs}} \sigma_{\log_{10}L_{\rm obs}}} \exp\left( -\frac{1}{2} \left[ \frac{(\log_{10}T_{\rm obs} - \log_{10}T_{i})^2}{\sigma_{\log_{10}T_{\rm obs}}^2} + \frac{(\log_{10}L_{\rm obs} - \log_{10}L_{i})^2}{\sigma_{\log_{10}L_{\rm obs}}^2} \right] \right), \tag{1}
+$$ -->
 $$
-\mathcal{L}(T_{i}, L_{i} | T_{\rm obs}, L_{\rm obs}) = \frac{1}{2\pi \sigma_{\log_{10}T_{\rm obs}} \sigma_{\log_{10}L_{\rm obs}}} \exp\left( -\frac{1}{2} \left[ \frac{(\log_{10}T_{\rm obs} - \log_{10}T_{i})^2}{\sigma_{\log_{10}T_{\rm obs}}^2} + \frac{(\log_{10}L_{\rm obs} - \log_{10}L_{i})^2}{\sigma_{\log_{10}L_{\rm obs}}^2} \right] \right), \tag{1}
+\begin{aligned}
+&\mathcal{L}(\log_{10}T_{i}, \log_{10}L_{i} \mid \log_{10}T_{\rm obs}, \log_{10}L_{\rm obs}) = 
+\frac{1}{2\pi \sigma_{\log_{10}T_{\rm obs}} \sigma_{\log_{10}L_{\rm obs}}} \\
+&\quad \times \exp\left( -\frac{1}{2} \left[ 
+\frac{(\log_{10}T_{\rm obs} - \log_{10}T_{i})^2}{\sigma_{\log_{10}T_{\rm obs}}^2} 
++ \frac{(\log_{10}L_{\rm obs} - \log_{10}L_{i})^2}{\sigma_{\log_{10}L_{\rm obs}}^2} 
+\right] \right),
+\end{aligned} \tag{1}
 $$
 where the $T$ and $L$ are the effective temperature and bolometric luminosity, respectively. $T_i, L_i$ are the values from the evolutionary model grids, and $T_{\rm obs}, L_{\rm obs}$ are the observed values for each target with their uncertainties ($1\,\sigma$ from the assumed Gaussian distribution **log-scale**) described as $\sigma_{\log_{10}T_{\rm obs}}, \sigma_{\log_{10}L_{\rm obs}}$.
[…]
