+++ paper/paper.md
@@ -38,8 +38,8 @@ $$
 \right] \right),
 \end{aligned} \tag{1}
 $$
-where the $T$ and $L$ are the effective temperature and bolometric luminosity, respectively. $T_i, L_i$ are the values from the evolutionary model grids, and $T_{\rm obs}, L_{\rm obs}$ are the observed values for each target with their uncertainties ($1\,\sigma$ from the assumed Gaussian distribution **log-scale**) described as $\sigma_{\log T_{\rm obs}}, \sigma_{\log L_{\rm obs}}$.
-**In this first released version, we assume log-uniform priors for $T$ and $L$ in this likelihood function following the `IDL` code used in @Pascucci_2016. This is because both initial mass function of stars and their evolutionary timescales imply that the occurrence of stars decreases as a function of $T_{\rm eff}$ and $L_{\rm bol}$, which is represented by the log-uniform priors. Different likelihood function can be added in the future versions.**
+where the $T$ and $L$ are the effective temperature and bolometric luminosity, respectively. $T_i, L_i$ are the values from the evolutionary model grids, and $T_{\rm obs}, L_{\rm obs}$ are the observed values for each target with their uncertainties ($1\,\sigma$ from the assumed Gaussian distribution **in log-scale**) described as $\sigma_{\log T_{\rm obs}}, \sigma_{\log L_{\rm obs}}$.
+**In this first released version, we follow the `IDL` code used in @Pascucci_2016, where we assume log-uniform priors for $T$ and $L$ in this likelihood function. This is because both initial mass function of stars and their evolutionary timescales imply that the occurrence of stars decreases as a function of $T_{\rm eff}$ and $L_{\rm bol}$. Different likelihood function with different priors can be added in the future versions.**
 
 **We choose $T_{\rm eff}$ and $L_{\rm bol}$ to estimate the stellar age and mass because extinction is significant for young stars especially when embedded in the natal cloud. Although the $T_{\rm eff}$ and $L_{\rm bol}$ are not directly observed quantities, they are the two main quantities that evolutionary models can be compared with. When median or high-resolution spectroscopy is employed on individual targets, $T_{\rm eff}$ and $L_{\rm bol}$ can be well determined, and the best estimates for YSOs are from works where a stellar spectrum is fitted simutaneously with extinction and accretional heating [e.g., @alcala_x-shooter_2017]. To ensure the best results, we recommend using $T_{\rm eff}$ and $L_{\rm bol}$ (with their uncertainties) that are derived simutaneously through spectroscopy.**
 
