### 2025-01-22 4fb6951
update paper, index and documentation
