### 2025-01-22 #7493 @DingshanDeng
Hi @HeloiseS and @vsquicciarini

Thanks for your helpful comments and suggestions. I have implemented them and you can find the updated texts in boldface in this new article proof. I attempted to solve the [issues](https://github.com/DingshanDeng/ysoisochrone/issues) @HeloiseS created, please find the reply there. Here I also provide a summary of replys on other comments below.
