### 2025-01-22 #7493 @DingshanDeng
Here are the comments provided by @HeloiseS , followed by the reply to them.

> -  YSOs should be expanded the first time (it's expanded the second time)
> -  line 35 and 36: looks like a list of bullet points didn't format properly
> -  line 44, the examples aren't showing up.

Re:
All of the comments above are addressed and the paper is updated. The updated texts in the paper are shown in boldface.

> A more general question for [@DingshanDeng](https://github.com/DingshanDeng): do none of the other track fitting algorithm allow you to use your own YSO tracks but with their fitting methods?
> For example isn't there other open source codes where you could instead add the YSO functionality rather than create a whole new package?
> I'd like a bit more text on why a separate code is needed and joining other endeavours isn't workable.

Re:
ysoisochrone package is based on the IDL code that was used in the literature but was not publicly available (e.g., Pascucci et al. 2016). A similar Bayesian method to derive stellar masses and ages for young stellar objects was adopted in many previous works (e.g., Andrews et al. 2013), but none of those codes was published. There are other tools/packages that could be used to estimate the stellar masses and ages, including the Python package isochrone (https://github.com/timothydmorton/isochrones), MADYS (https://madys.readthedocs.io/en/latest/) as well as a web interface to use the PARSEC tracks (http://stev.oapd.inaf.it/PARSEC/tools.html). However, none of them provides easy access to using the Bayesian framework together with different pre-main-sequence stellar evolutionary models. In addition, the MADYS code uses photometry to estimate the bolometric luminosity and temperature, which is good for oldish/non-embedded sources but problematic for the majority of YSOs (see also comments and replies to @vsquicciarini below). Here we publish this package to give easy access to using the published isochrones and the Bayesian method to people in our community. This is discussed and clarified in the statement of need section.
