### 2025-01-22 0bf99a2
update paper, index and documentation
