### 2025-01-22 07236d2
update paper, index and documentation
