### 2025-01-22 #7493 @DingshanDeng
Here are the comments provided by @vsquicciarini , I reordered them a little bit, putting the major concers at top. The replies follow each comments/questions.

## On the choice of input observables

> As I mentioned before, I think it is necessary to explain why (L, Teff) were chosen over photometry as input variables. Based on my experience, I consider the latter to be much more reliable than the former because:
> luminosity and effective temperatures are not observed quantities, but they are derived from photometry assuming some atmospheric model. Photometry is a real observable, only dependent on the filter response and the zero point of the instrument;
> being model-dependent, two measurements of L and Teff from two different papers might assume different underlying models, ages, parallaxes, metallicities, bolometric corrections, and be based on different observational data (photometry, spectroscopy); this makes any direct comparison of input data extremely difficult;
> as a consequence of 1) and 2), a (L, Teff) point obtained through a model A and then fitted by your package to model B will be affected by a systematic error which is not just a random error, but also comprises the systematic difference between model A and model B. How is this taken into account in the error estimation?

Re:
It is true that Lbol and Teff are not directly observed quantities but they are the two main quantities that evolutionary models can be compared with. For young stars, extinction is significant: unfortunately, we cannot use the observed photometry and colors without accounting for the extinction for young stars which varies across the region they belong to. When medium or high-resolution spectroscopy is employed on individual targets, one can better determine  Lbol and Teff. Therefore, the best estimates for the young stellar objects in nearby star-forming regions are from works where a stellar spectrum is fit simultaneously with extinction and accretional heating (e.g, Alcalá et al. 2017 https://ui.adsabs.harvard.edu/abs/2017A%26A...600A..20A/abstract). For point (2), the results of Teff and L that are determined simultaneously from the models are preferred. These more accurate Lbol and Teff are available for most nearby star-forming regions, see e.g., Manara et al. (2023) (https://ui.adsabs.harvard.edu/abs/2023ASPC..534..539M/abstract). On point (3), we clearly state in the main page of the documentation (https://ysoisochrone.readthedocs.io/en/latest/index.html) that users should ensure that Teff and Lbol are derived simultaneously through spectroscopy. As a note, these derivations do not require the use of an evolutionary mode. In some instances, e.g. Alcalá et al. 2017 (https://ui.adsabs.harvard.edu/abs/2017A%26A...600A..20A/abstract), spectral models are adopted but in other cases, e.g. Fang+2023 (https://ui.adsabs.harvard.edu/abs/2023ApJ...945..112F/abstract), the derivation of Teff and Lbol is completely empirical by using spectra of non-accreting young stars.

## Comments on the paper

>  line 9: stellar age -> stellar ages.

Re: the text in paper is updated.

>  line 15: I would use the singular: the stellar mass, age, its effective temperature etc.

Re: the text in paper is updated.

> line 18: please add references here for "as assumed in previous literature". The point here is not trivial: to me, a log-uniform prior for Teff and L is more physically sound than a uniform prior. The assumption of linearly uniform priors is not uninformative, because it creates a bias for hotter and brighter stars. Both the shape of the initial mass function of stars and their evolutionary timescales imply that the occurrence of stars decreases as a function of Teff and L. Hence, using a log-uniform prior would give a more unbiased estimates of the underlying distribution. I know that this is a strong request, but I would suggest replacing the observables Teff and L with logTeff and logL.

Re: We add the citations to the text. The referee is correct, using logTeff and logL with log-uniform prior is a better choice that creates less bias. Indeed, in both the IDL version (used in Pascucci et al. 2016) of the code and this current version, we calculate the likelihood function with logTeff and logLbol. We clarify this in the text in line 19-27 and updated equation (1).

> line 24: it should be specified that the Feiden models here are non-magnetic. I could find this information from the quoted Pascucci+16 reference, but I think that this should be explicitly stated.

Re: We add that information in line 36.

> line 24: the choice of combining Baraffe and Feiden models is based on the study by Herczeg & Hillenbrand (2015), quoted in Pascucci+16. While I do not question their models and results, I note that over the last 10 years there were both significant improvements of stellar models and the observational revolution brought about by Gaia. My point is that a more recent study should be used here to justify the choice of the isochrone set. If this is not possible, I think that it would be perfectly fine to state more explicitly that your choice is based on the 2015 analysis, that might not be up to date, and that you intend to add new models in later versions of your package.

Re: The combination of Baraffe and Feiden was initially suggested in Herczeg & Hillenbrand (2015), and was adopted later in many works, including the most recent review chapter of Manara et al. (2023). There has been some recent development of the models, such as the recent Feiden models (as mentioned in their webpage https://github.com/gfeiden/MagneticUpperSco/tree/master/models). However, the updated tracks are not publicly available at this point, and therefore we still adopt the non-magnetic tracks in this open-source package. This is clarified in line 38-41.

> general comment: I would appreciate if you could detail in the text why you chose to use as input parameters L and Teff, instead of the measured photometry. On my side, I find this choice problematic for several reasons (that I will detail below); but I am not really an expert of proto-stars and <5 Myr-old stars with a full disk, so there might be reasons for your choice that I am overlooking.

Re: As we replied above, the combination of Teff and Lbol could give better estimates of stellar age and mass for those young stellar objects in nearby star-forming regions. We add that information in line 28-35.

> statement of need: I think this has to be better rephrased: the e.g. parethesis is currently empty, and it is not clear what the advantage of this package over previous work is. I'm sorry to say that, but I have written a fully documented and open source package, MADYS (https://madys.readthedocs.io/en/latest/), that can be used to derive ages and masses for pre-main sequence stars. Another package, isochrones, is cited in the documentation but not here. I think that both packages should be, if not discussed, at least acknowledged here.

Re: Thanks for pointing out these published tools. We add more detailed discussions and acknowledgement in the statement of need section, as well as in the documentation and main page on github.

## comments on the Jupyter Notebook

> As I mentioned above, I may write some additional technical comments on the code itself after testing it more carefully. For the moment being, I just write here some general comments that came to my mind while executing the notebook:
>  4.3) Estimate mass and ages: I understand the practical advantage of attempting to estimate masses even for underluminous stars. However, I think that the results here should be taken with extreme caution, and the user should be aware of that. I would add a flag column in df_output that specifies if the star is within the grid, underluminous or overluminous. Also, I notice that no error is provided on the mass and age here. This should be fixed, because 1) uncertainties must always be estimated, 2) for non-standard stars is it even more crucial to understand how much the mass is constrained. The uncertainty on Teff and the uncertainty on the assume age should be used to evaluate the corresponding mass uncertainty.
>  5.1) Closest grid point on isochrone: as above, I think it is mandatory to provide uncertainty estimation here. I would estimate the uncertainties in a Monte Carlo way, i.e. by picking the closest (mass, age) grid point but with L and Teff randomly altered N times (let's say, 1000) according to their uncertainties. While I understand that this is not the preferred method and it is only provided to reproduce literature results, I stress the importance of doing that to allow for meaningful comparison with the literature itself.
>  5.2) Assuming age to derive stellar masses: it is not clear to me what the difference between the two possibilities is. Am I right in thinking that in the first case it interpolates across mass, while in the second just takes the closest track? I think this should be better rephrased.

Re:
We add a new column in the df_output called ‘flag’ to specify whether the estimate is ‘good’, or the star is ‘toobright’ or ‘toofaint’. We do not provide the uncertainty for age because it is an assumed value. The uncertainties on stellar mass estimates are also added for those targets. On the other points, the referee is correct. In both Bayesian and the closest track approach, the code utilizes the interpolation to create a meshgrid based on the input stellar evolutionary tracks and then find the mass. We have clarified in the documentation, this code provides the Bayesian analysis (that is preferred and could provide uncertainties), and has a stand alone function to find the closest grid points to test the results against literature (e.g., Pascucci et al. 2016, Manara et al. 2023). We agree with the referee, this method that simply finds the closest tracks should not be used, and we recommend users to use the Bayesian approach, where the uncertainties are provided, this is clearly mentioned in the updated version of paper, documentation and tutorial notebooks.

> general comment: if a star is underluminous due to disk absorption, we should expect an effect on temperature as well. This means that when we use an age and a Teff, the Teff should be biased to lower values too (due to differential absorption). Shouldn't this be taken into account by using some kind of de-reddening? Also, depending on the way Teff was derived, Teff itself might be more or less biased. For instance, if Teff was derived only using NIR data, then the disk luminosity would introduce a larger bias compared to cases where UV measurements are available. Again, this stresses how the choice of Teff as an input parameter can induce a lot of unseen systematics.

Re:
Please see the reply above on why we choose to use the Teff and Lbol as input parameters.

## Comments on the documentation

>  "The available stellar evolutionary models": in the sentence "This package is mainly built to study the young-stellar-objects, so by default; it only utilizes the stellar evolutionary tracks for pre-main-sequence stars ({\rm stellar\ age}< 500,{\rm Myrs}). " , I think the parenthesis is misleading because the duration of the pre-MS strongly depends on mass. A solar mass star only stays on the pre-MS for about 50 Myr. I suggest removing the parenthesis.

Re: it is removed.

>  In general, I find that the paragraph above could be rewritten better. I suggest: "For the moment being, this package is meant to be used to study young stellar objects. In other words, it only utilizes the stellar evolutionary tracks for pre-main-sequence stars, avoiding the problem of dealing with main-sequence and post-main-sequence targets (whose luminosity rises again and will overlay on the pre-main-sequence phase). In future releases, the package might be expanded to also handle main-sequence stars (e.g., Fernandes et al. 2023), but user-specified customized stellar evolutionary tracks would be needed."

Re: this paragraph updated accordingly.

>  Feiden: again, it should be specified that the isochrones are non-magnetic.

Re: it is specified there.

>  MIST: since you are quoting the isochrones package, I suggest quoting my MADYS package as well, which can handle MIST, PARSEC (1.2 and 2.0), Feiden, Baraffe and many other models for pre-MS or MS stars.

Re: MADYS is quoted and discussed in the documentation, as well as in the paper and main page on github.
