### 2025-01-23 #7493 @vsquicciarini
Dear @DingshanDeng, thanks a lot for considering my comments and taking time to implement or tackle them. I really appreciate that, and I can see that you carried out the endeavor with plenty of zeal. I have labeled all of them as solved. I have reread the manuscript and I don't have any additional comment to make, apart from a couple of typos ("from the stellar evolutionary models" ->  "from stellar evolutionary models" at line 9, and "median or high-resolution spectroscopy" -> "medium or high-resolution spectroscopy").

In addition to this, I am grateful for the clarification regarding the choice of (L, Teff) as inputs. Indeed extinction is a huge problem when dealing with photometry, and a joint fit of (L, Teff, extinction) from spectroscopic data can be a game changer when the target is < 5 Myr old. This is now adequately explained in the text.

As a final check, I am going to rerun the notebooks and (possibly) add a few comments regarding the code. I will promptly inform you when my tests are over.
