### 2025-01-23 #7493 @vsquicciarini
A couple of comments related to the function `ysoisochrone.plotting.plot_hr_diagram()`.

I think that the legend here can be improved. As all tracks (isochrones) are plotted with the same line style and color, adding an individual label for each track (isochrone) in the legend is unnecessary. Besides decreasing the elegance of the plot – otherwise really cool! –  by taking space without adding scientific value, this can also induce problems if a user plots more tracks and isochrones than the figure height can include. An example is provided in the first figure here:

![Image](https://github.com/user-attachments/assets/c2a12f1a-7766-4fe2-a7c2-e23a17a60ed4)

I suggest putting a single label for isochrones and a single label for tracks in this way:

![Image](https://github.com/user-attachments/assets/f05f0011-1bda-4731-8474-80eafe744209)

or removing the labels for tracks and isochrones altogether.

My second remark is that there is a bug when setting the keyword `bare` to `True`. Labels are only removed from the right plot.

![Image](https://github.com/user-attachments/assets/ca6597d3-1115-469c-9bae-758a1a2bb57d)
