### 2025-01-23 #7493 @HeloiseS
Hi @DingshanDeng!

Thanks for making the suggested changes and adding text to explain the purpose and need for the code.
Have you updated the pip version? I can see it's dated form Jan 15th but the unittests are form last week and the notebooks were updated a few hours ago
