### 2025-01-23 #7493 @vsquicciarini
As a general question, to my eyes it seems from the previous plots that the two theoretical grids, albeit very close, do not exactly overlap. Is that true?
* If the models do not overlap, I think it has to be stated explicitly in the documentation, providing a figure like Fig. 1 of Pascucci+16. In this case, I expect the situation for a star lying close to T_eff=3900 K to be tricky. If the star is fitted with Feiden isochrones if T_eff > 3900 K and with Baraffe isochrones otherwise, the shape of the posteriors will be skewed (I mean, it should become non-continuous and bimodal). You may want to investigate the impact of this on the derived parameters and uncertainties.
* If the models are simultaneously fit with a polynomial curve to merge them into a single model, the process needs to be explained in the documentation.
