### 2025-01-23 #7493 @DingshanDeng
> Hi [@DingshanDeng](https://github.com/DingshanDeng)!
>
> Thanks for making the suggested changes and adding text to explain the purpose and need for the code. Have you updated the pip version? I can see it's dated form Jan 15th but the unittests are form last week and the notebooks were updated a few hours ago. (once that's done I'll update your code on my conda env, and rerun all the tests and notebooks - if it all runs that'll be it on my end!)

Thanks for the reminder. I have not updated the pip version. I will update that once I finished the tests/comments suggested above :-).
