### 2025-02-02 fc1ed20
update paper and tutorials; add new feature for plotting; fixed a bug in the isochrone
