+++ paper/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 ---
 
 # Background and Methods
-There has been a long history of estimating **stellar ages and masses** from the stellar evolutionary models [e.g., @Baraffe_2015; @Feiden_2016; @Siess_2000]. Different methods have been employed, from finding the closest track to an object's luminosity and temperature [e.g., @Manara_2023_PPVII] to employing a Bayesian approach which enables estimating uncertainties on the inferred ages and masses [e.g., @Andrews_2013; @Gennaro_2012;  @Jorgensen_n_Lindergren_2005]. Our primary method is a Bayesian inference approach, and the `Python` code builds on the `IDL` version developed by @Pascucci_2016. **The code estimates the stellar mass, age, and associated uncertainties by comparing a star's effective temperature ($T_{\rm eff}$), bolometric luminosity ($L_{\rm bol}$)**, and their uncertainties with different stellar evolutionary models, including those specifically developed for young stellar objects **(YSOs)**. **The conditional likelihood function assumes log-uniform priors and can be written as:**
+There has been a long history of estimating stellar ages and masses from the stellar evolutionary models [e.g., @Baraffe_2015; @Feiden_2016; @Siess_2000]. Different methods have been employed, from finding the closest track to an object's luminosity and temperature [e.g., @Manara_2023_PPVII] to employing a Bayesian approach which enables estimating uncertainties on the inferred ages and masses [e.g., @Andrews_2013; @Gennaro_2012;  @Jorgensen_n_Lindergren_2005]. Our primary method is a Bayesian inference approach, and the `Python` code builds on the `IDL` version developed by @Pascucci_2016. The code estimates the stellar mass, age, and associated uncertainties by comparing a star's effective temperature ($T_{\rm eff}$), bolometric luminosity ($L_{\rm bol}$), and their uncertainties with different stellar evolutionary models, including those specifically developed for young stellar objects (YSOs). The conditional likelihood function assumes log-uniform priors and can be written as:
 $$
 \begin{aligned}
 &\mathcal{L}(\log T_{i}, \log L_{i} \mid \log T_{\rm obs}, \log L_{\rm obs}) = 
@@ -38,12 +38,12 @@ $$
 \right] \right),
 \end{aligned} \tag{1}
 $$
-where the $T$ and $L$ are the effective temperature and bolometric luminosity, respectively. $T_i, L_i$ are the values from the evolutionary model grids, and $T_{\rm obs}, L_{\rm obs}$ are the observed values for each target with their uncertainties ($1\,\sigma$ from the assumed Gaussian distribution **in log-scale**) described as $\sigma_{\log T_{\rm obs}}, \sigma_{\log L_{\rm obs}}$.
-**In this first released version, we follow the `IDL` code used in @Pascucci_2016, where we assume log-uniform priors for $T$ and $L$ in this likelihood function. This is because both initial mass function of stars and their evolutionary timescales imply that the occurrence of stars decreases as a function of $T_{\rm eff}$ and $L_{\rm bol}$. Different likelihood function with different priors can be added in the future versions.**
+where the $T$ and $L$ are the effective temperature and bolometric luminosity, respectively. $T_i, L_i$ are the values from the evolutionary model grids, and $T_{\rm obs}, L_{\rm obs}$ are the observed values for each target with their uncertainties ($1\,\sigma$ from the assumed Gaussian distribution in log-scale) described as $\sigma_{\log T_{\rm obs}}, \sigma_{\log L_{\rm obs}}$.
[…]
