+++ ysoisochrone/plotting.py
@@ -35,8 +35,9 @@ style = [
 plt.style.use(style)
 
 from ysoisochrone import utils
+from ysoisochrone import isochrone
 
-def plot_bayesian_results(log_age_dummy, log_masses_dummy, L, best_age, best_mass, age_unc, mass_unc, source=None, save_fig=False, fig_save_dir='figure', customized_fig_name=''):
+def plot_bayesian_results(log_age_dummy, log_masses_dummy, L, best_age, best_mass, age_unc, mass_unc, source=None, save_fig=False, fig_save_dir='figure', customized_fig_name='', color_bestfit='red', color_likelihood='blue', cmap_likelihood='viridis'):
     """
     Plots the likelihood distributions and the best-fit age and mass.
 
@@ -64,6 +65,12 @@ def plot_bayesian_results(log_age_dummy, log_masses_dummy, L, best_age, best_mas
             Directory to save the figure if save_fig is True.
         customized_fig_name [str, optional]: 
             Customized figure name.
+        color_bestfit [str, optional]:
+            Customized color to show the best fit point and uncertainties
+        color_likelihood [str, optional]:
+            Customized color to show the the likelihood function
+        cmap_likelihood [str, optional]:
+            Customized color to show the the likelihood cmap
         
     Output:
     
@@ -141,30 +148,34 @@ def plot_bayesian_results(log_age_dummy, log_masses_dummy, L, best_age, best_mas
     ax_marg_y.xaxis.grid(False)
     
     # Plot marginal age likelihood
-    ax_marg_x.plot(log_age_dummy, np.sum(L, axis=1), color='blue', linestyle='-', marker='o')
-    ax_marg_x.axvline(best_age, color='red', linestyle='-')
-    ax_marg_x.axvline(age_unc[0], color='red', linestyle='--')
-    ax_marg_x.axvline(age_unc[1], color='red', linestyle='--')
+    ax_marg_x.plot(log_age_dummy, np.sum(L, axis=1), color=color_likelihood, linestyle='-', marker='o')
+    ax_marg_x.set_xlim(log_age_dummy.min(), log_age_dummy.max())
+    ax_marg_x.axvline(best_age, color=color_bestfit, linestyle='-')
+    ax_marg_x.axvline(age_unc[0], color=color_bestfit, linestyle='--')
+    ax_marg_x.axvline(age_unc[1], color=color_bestfit, linestyle='--')
     ax_marg_x.set_ylabel('Likelihood')
 
     # Plot marginal mass likelihood
-    ax_marg_y.plot(np.sum(L, axis=0), log_masses_dummy, color='blue', linestyle='-', marker='o')
-    ax_marg_y.axhline(best_mass, color='red', linestyle='-')
-    ax_marg_y.axhline(mass_unc[0], color='red', linestyle='--')
-    ax_marg_y.axhline(mass_unc[1], color='red', linestyle='--')
+    ax_marg_y.plot(np.sum(L, axis=0), log_masses_dummy, color=color_likelihood, linestyle='-', marker='o')
+    ax_marg_y.set_ylim(log_masses_dummy.min(), log_masses_dummy.max())
+    ax_marg_y.axhline(best_mass, color=color_bestfit, linestyle='-')
+    ax_marg_y.axhline(mass_unc[0], color=color_bestfit, linestyle='--')
+    ax_marg_y.axhline(mass_unc[1], color=color_bestfit, linestyle='--')
     ax_marg_y.set_xlabel('Likelihood')
 
     # Prepare data for the joint likelihood plot
     data = copy.deepcopy(L)
-    data[np.isnan(data)] = 1e-99
[…]

[…]
