+++ ysoisochrone/__init__.py
@@ -12,7 +12,7 @@ from . import utils
 from . import plotting
 from . import isochrone
 
-__version__ = "0.4.0.beta"
+__version__ = "0.4.1.beta"
 __author__ = "Dingshan Deng"
 __copyright__ = "Copyright (C) 2024 Dingshan Deng"
 __all__ = ["isochrone", "bayesian", "plotting", "utils"]
