### 2025-02-03 24b698a
update paper and documentation
