+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "ysoisochrone"
-version = "0.4.1.beta"
+version = "1.0.0"
 authors = [
   { name="Dingshan Deng", email="dingshandeng@arizona.edu" },
 ]
