### 2025-02-10 #7493 @warrickball
I'd just like to check in with @HeloiseS, who [wrote](https://github.com/openjournals/joss-reviews/issues/7493#issuecomment-2609565940)

> Thanks for making the suggested changes and adding text to explain the purpose and need for the code. Have you updated the pip version? I can see it's dated form Jan 15th but the unittests are form last week and the notebooks were updated a few hours ago. (once that's done I'll update your code on my conda env, and rerun all the tests and notebooks - if it all runs that'll be it on my end!)

I see that you've ticked everything off in the review checklist. Have you run everything you wanted to and, if so, can you confirm you're happy to recommend the paper be accepted? Thanks!
