### 2025-02-10 #7493 @warrickball
Great! In that case, @DingshanDeng, the paper is accepted. Congratulations! I'll shortly generate a post-review checklist of items to work through. The main one will be for you to create an archive of the code to upload to an archiving service like Zenodo or Figshare.

I'll also start doing some editorial checks of the paper and either post comments here or open a PR for the finer details. I probably won't get to any of this tonight, though!
