### 2025-02-10 #7493 @DingshanDeng
We double checked the authors and affiliations including ORCIDs. We made a release with the latest changes, and it is release v1.0.0. (or v1.0.0-zenodo, which is the same). We also archiev the release on Zenodo, and the reference is here: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14847202.svg)](https://doi.org/10.5281/zenodo.14847202) with the DOI number: `10.5281/zenodo.14847202`. The title, author list and the license in the archive matches in the JOSS paper. Thanks!
