### 2025-02-12 #7493 @warrickball
Hi @DingshanDeng, thanks for creating the archive, where everything looks correct. I'll have a closer look at the paper hopefully over the next few days but there are a few things that you can address already.

1. I think it makes more sense to put the *Statement of Need* at the start of the paper, as it leads better than the current *Backgrounds and Methods* section.
2. You've presumably gathered many of your citations from NASA ADS, which uses shorthands like `\mnras` for journal names and our pipeline doesn't understand. Could you replace these in the BibTeX with the full journal names? E.g. *Monthly Notices of the Royal Astronomical Society*. (This is very common in astronomy submissions to JOSS.)
3. There are two citations to arXiv versions of articles that have published versions, specifically [Siess et al. (2000)](https://ui.adsabs.harvard.edu/abs/2000A%26A...358..593S) and [Manara et al. (2023)](https://ui.adsabs.harvard.edu/abs/2023ASPC..534..539M). Please update these to the published, not arXiv, versions.

I think this will already takes us very close to publication, so after these are merged I'll probably open a pull request to address minor editorial points (if there are any).
