+++ paper/paper.bib
@@ -39,7 +39,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Bressan_2012_PARSEC,
        author = {{Bressan}, Alessandro and {Marigo}, Paola and {Girardi}, L{\'e}o. and {Salasnich}, Bernardo and {Dal Cero}, Claudia and {Rubele}, Stefano and {Nanni}, Ambra},
         title = "{PARSEC: stellar tracks and isochrones with the PAdova and TRieste Stellar Evolution Code}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {stars: evolution, Hertzsprung{\ensuremath{-}}Russell and colour magnitude diagrams, stars: interiors, stars: low-mass, Astrophysics - Solar and Stellar Astrophysics},
          year = 2012,
         month = nov,
@@ -58,7 +58,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Siess_2000,
        author = {{Siess}, L. and {Dufour}, E. and {Forestini}, M.},
         title = "{An internet server for pre-main sequence tracks of low- and intermediate-mass stars}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {EQUATION OF STATE, STARS: EVOLUTION, STARS: HERTZSPRUNG-RUSSEL (HR) AND C-M DIAGRAMS, STARS: PRE-MAIN SEQUENCE, ASTRONOMICAL DATA BASES: MISCELLANEOUS, Astrophysics},
          year = 2000,
         month = jun,
@@ -116,7 +116,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Gennaro_2012,
        author = {{Gennaro}, M. and {Prada Moroni}, P.~G. and {Tognelli}, E.},
         title = "{Testing pre-main-sequence models: the power of a Bayesian approach}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {methods: statistical, binaries: eclipsing, binaries: general, stars: fundamental parameters, stars: pre-main-sequence, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Astrophysics of Galaxies, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2012,
         month = feb,
@@ -135,7 +135,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Jorgensen_n_Lindergren_2005,
        author = {{J{\o}rgensen}, B.~R. and {Lindegren}, L.},
         title = "{Determination of stellar ages from isochrones: Bayesian estimation versus isochrone fitting}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {stars: fundamental parameters, stars: evolution, solar neighbourhood, methods: data analysis, methods: statistical},
          year = 2005,
         month = jun,
@@ -171,7 +171,7 @@ ADS Bibcode: 2015ApJ...808...23H},
 @ARTICLE{Pascucci_2016,
        author = {{Pascucci}, I. and {Testi}, L. and {Herczeg}, G.~J. and {Long}, F. and {Manara}, C.~F. and {Hendler}, N. and {Mulders}, G.~D. and {Krijt}, S. and {Ciesla}, F. and {Henning}, Th. and {Mohanty}, S. and {Drabek-Maunder}, E. and {Apai}, D. and {Sz{\H{u}}cs}, L. and {Sacco}, G. and {Olofsson}, J.},
         title = "{A Steeper than Linear Disk Mass-Stellar Mass Scaling Relation}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
[…]
