+++ paper/paper.bib
@@ -64,7 +64,6 @@ archivePrefix = {arXiv},
         month = jun,
        volume = {358},
         pages = {593-599},
-          doi = {10.48550/arXiv.astro-ph/0003477},
 archivePrefix = {arXiv},
        eprint = {astro-ph/0003477},
  primaryClass = {astro-ph},
@@ -104,12 +103,12 @@ ADS Bibcode: 2019ApJ...884...42S},
        volume = {534},
         month = jul,
         pages = {539},
-          doi = {10.48550/arXiv.2203.09930},
 archivePrefix = {arXiv},
        eprint = {2203.09930},
  primaryClass = {astro-ph.SR},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2023ASPC..534..539M},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+          utl = {https://par.nsf.gov/biblio/10385775.}
 }
 
 
