+++ paper/paper.bib
@@ -64,11 +64,9 @@ archivePrefix = {arXiv},
         month = jun,
        volume = {358},
         pages = {593-599},
-archivePrefix = {arXiv},
-       eprint = {astro-ph/0003477},
- primaryClass = {astro-ph},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2000A&A...358..593S},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+          url = {https://adsabs.harvard.edu/full/2000A%26A...358..593S}
 }
 
 
@@ -103,12 +101,9 @@ ADS Bibcode: 2019ApJ...884...42S},
        volume = {534},
         month = jul,
         pages = {539},
-archivePrefix = {arXiv},
-       eprint = {2203.09930},
- primaryClass = {astro-ph.SR},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2023ASPC..534..539M},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System},
-          utl = {https://par.nsf.gov/biblio/10385775.}
+          utl = {https://www.aspbooks.org/a/volumes/article_details/?paper_id=40759}
 }
 
 
