+++ paper/paper.bib
@@ -101,9 +101,7 @@ ADS Bibcode: 2019ApJ...884...42S},
        volume = {534},
         month = jul,
         pages = {539},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2023ASPC..534..539M},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System},
-          utl = {https://www.aspbooks.org/a/volumes/article_details/?paper_id=40759}
+          url = {https://www.aspbooks.org/a/volumes/article_details/?paper_id=40759}
 }
 
 
