### 2025-02-12 #7493 @DingshanDeng
> Hi [@DingshanDeng](https://github.com/DingshanDeng), thanks for creating the archive, where everything looks correct. I'll have a closer look at the paper hopefully over the next few days but there are a few things that you can address already.
>
> 1. I think it makes more sense to put the _Statement of Need_ at the start of the paper, as it leads better than the current _Backgrounds and Methods_ section.
> 2. You've presumably gathered many of your citations from NASA ADS, which uses shorthands like `\mnras` for journal names and our pipeline doesn't understand. Could you replace these in the BibTeX with the full journal names? E.g. _Monthly Notices of the Royal Astronomical Society_. (This is very common in astronomy submissions to JOSS.)
> 3. There are two citations to arXiv versions of articles that have published versions, specifically [Siess et al. (2000)](https://ui.adsabs.harvard.edu/abs/2000A%26A...358..593S) and [Manara et al. (2023)](https://ui.adsabs.harvard.edu/abs/2023ASPC..534..539M). Please update these to the published, not arXiv, versions.
>
> I think this will already takes us very close to publication, so after these are merged I'll probably open a pull request to address minor editorial points (if there are any).

Hi @warrickball, thanks for your suggestions and to review this.

I updated the draft and I moved the statement of need to the start of the paper.

I also updated the journal names in the bib file and now it should look good in the pdf.

For the two citations, I could not find their DOI numbers, and it might because the  [Siess et al. (2000)](https://ui.adsabs.harvard.edu/abs/2000A%26A...358..593S) might be too old to have an assigned DOI number while the [Manara et al. (2023)](https://ui.adsabs.harvard.edu/abs/2023ASPC..534..539M) is a book. So I put the url links for the published versions.
