### 2025-02-12 #7493 @warrickball
Thanks @DingshanDeng! I'll have a look over the next few days.

> For the two citations, I could not find their DOI numbers, and it might because the [Siess et al. (2000)](https://ui.adsabs.harvard.edu/abs/2000A%26A...358..593S) might be too old to have an assigned DOI number while the [Manara et al. (2023)](https://ui.adsabs.harvard.edu/abs/2023ASPC..534..539M) is a book. So I put the url links for the published versions.

Though we like to have a complete set of DOIs, we are aware that there can be gaps. Both the reasons you mention are quite common and perfectly reasonable.
