### 2025-02-16 751ae7a
Initial Editorial tweaks for JOSS
