+++ paper/paper.md
@@ -12,51 +12,49 @@ authors:
     affiliation: 1
   - name: Rachel B. Fernandes
     orcid: 0000-0002-3853-7327
-    affiliation: "2, 3, 4"
+    affiliation: "2, 3"
 affiliations:
  - name: Lunar and Planetary Laboratory, the University of Arizona, Tucson, AZ 85721, USA
    index: 1
- - name: President’s Postdoctoral Fellow
+ - name: President's Postdoctoral Fellow, Department of Astronomy \& Astrophysics, 525 Davey Laboratory, The Pennsylvania State University, University Park, PA 16802, USA
    index: 2
- - name: Department of Astronomy \& Astrophysics, 525 Davey Laboratory, The Pennsylvania State University, University Park, PA 16802, USA
-   index: 3
  - name: Center for Exoplanets and Habitable Worlds, 525 Davey Laboratory, The Pennsylvania State University, University Park, PA 16802, USA
-   index: 4
+   index: 3
 date: 10 February 2025
 bibliography: paper.bib
 ---
 
 # Statement of Need
 
-At this moment, there are a few existing tools and packages that can be used to handle stellar evolutionary tracks and to estimate stellar mass and age for pre-main sequence stars. For example, [`isochrones`](https://github.com/timothydmorton/isochrones) provides a `Python` interface to access the [MIST](https://waps.cfa.harvard.edu/MIST/) grids, and the `PARSEC` team provides a [web interface](http://stev.oapd.inaf.it/PARSEC/tools.html) to access different versions of their tracks together with some useful web-based tools. More recently, @squicciarini_madys_2022 developed and published another `Python` package [`MADYS`](https://madys.readthedocs.io/en/latest/), which can be used to derive ages and masses for pre-main sequence stars from multi-wavelengths photometric data with the extinction corrected according to extinction maps and laws. This code could also utilize different stellar evolutionary models. `MADYS` provides easy access to obtaining photometric age and mass estimates for large groups of young stellar or substellar objects.
+At this moment, there are a few existing tools and packages that can be used to handle stellar evolutionary tracks and to estimate stellar mass and age for pre-main sequence stars. For example, [`isochrones`](https://github.com/timothydmorton/isochrones) [@isochrones] provides a Python interface to access the [MIST](https://waps.cfa.harvard.edu/MIST/) grids [@Dotter_2016; @Choi_2016], and the PARSEC team provides a [web interface](http://stev.oapd.inaf.it/PARSEC/tools.html) to access different versions of their tracks together with some useful web-based tools. More recently, @squicciarini_madys_2022 developed and published another Python package [`MADYS`](https://madys.readthedocs.io/en/latest/), which can be used to derive ages and masses for pre-main sequence stars from multi-wavelengths photometric data with the extinction corrected according to extinction maps and laws. This code could also utilize different stellar evolutionary models. `MADYS` provides easy access to obtaining photometric age and mass estimates for large groups of young stellar or substellar objects.
 
-Here we introduce `ysoisochrone`, a `Python3` package that utilizes stellar evolutionary tracks to estimate stellar masses and ages of pre-main sequence stars with a Bayesian framework. While several papers in the literature utilize this method [e.g., @Fernandes_2023; @Jorgensen_n_Lindergren_2005; @Pascucci_2016], an open-source tool implementing this method is not available. `ysoisochrone` fills this gap and provides a uniform platform to handle different evolutionary models with easy access to Bayesian framework along with tutorials and detailed documentation for first users. 
[…]
