### 2025-02-16 #7493 @DingshanDeng
> I've opened a [pull request](https://github.com/DingshanDeng/ysoisochrone/pull/8) in the `ysoisochrone` repo with some editorial tweaks. Once that's merged, I'll test the publication process, which might turn up a second round of editorial tweaks, but we'll have to see. We should be able to finish this up within a few days.

The pull request has now been merged into the base branch. Thanks!
