### 2025-02-17 #7493 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
