### 2025-02-17 #7493 @warrickball
Hi @DingshanDeng, though the message isn't clearly directed at anyone in particular, only (A)EiCs can run the `accept` command. I'll check the proof metadata deposits later today.
