### 2025-02-17 #7493 @warrickball
Thanks once again to @HeloiseS and @vsquicciarini for reviewing this submission! JOSS simply wouldn't be possible without its community of volunteers.
